Make module for optimism tx rpc response
